+++ stories/index.js
@@ -94,7 +94,7 @@ import './TableActionCell/TableActionCell.story'; // 10.1 Table -> TableActionCe
 import './DropdownLayout'; // 11.1 DropdownLayout
 import './EditableSelector'; // 11.2 EditableSelector
 import './ModalSelectorLayout.story.js'; // 11.3 ModalSelectorLayout
-import './ColorPicker.story.js'; // 11.5 Color Picker
+import './ColorPicker/index.story.js'; // 11.5 Color Picker
 
 // 12. Other
 import './Badge/index.story'; // 12.1 Badge
@@ -118,6 +118,7 @@ import './Text/index.story';
 import './Skeleton.story.js';
 import './Collapse.story';
 import './Popover/index.story';
+import './DropdownPopover/index.story';
 import './CardGalleryItem/CardGalleryItem.story';
 import './Carousel/index.story.js';
 import './CalendarPanel/index.story';